Create README.md
TILT focuses on bridging what we believe to be a major gap between merchant transactions and consumers. The crypto market is largely consumer-based and marginally accepted by merchants. There's a variety of reasons as to why this is. Throughout the release of our Whitepaper, coming out in Q1 of 2022, we hope to provide clarity on how we provide payment solutions for consumers that all merchants will accept without needing to change their current payment processors.